Fix float division
Firstly, this is intended to be a float division.

Secondly, dividing integers by 0 results in an exception, but dividing
non-zero floats by 0 results in +/- infinity which will be clamped to
the upper range.
In particular, this occurs when the beatmap has 1 hitobject (0 drain
length).

+++ osu.Game.Rulesets.Osu/Difficulty/OsuScoreV1Processor.cs
@@ -81,7 +81,7 @@ namespace osu.Game.Rulesets.Osu.Difficulty
                 (baseBeatmap.Difficulty.DrainRate
                  + baseBeatmap.Difficulty.OverallDifficulty
                  + baseBeatmap.Difficulty.CircleSize
-                 + Math.Clamp(objectCount / drainLength * 8, 0, 16)) / 38 * 5);
+                 + Math.Clamp((float)objectCount / drainLength * 8, 0, 16)) / 38 * 5);
 
             scoreMultiplier = difficultyPeppyStars * mods.Aggregate(1.0, (current, mod) => current * mod.ScoreMultiplier);
